requirements.txt
streamlit==1.30.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.11.4
fasttext==0.9.2
google-generativeai==0.4.0
PyYAML==6.0.1
python-dotenv==1.0.1
